% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_agg.R
\name{aggregate_elcitations}
\alias{aggregate_elcitations}
\title{Pool elcitations data}
\usage{
aggregate_elcitations(elcs)
}
\arguments{
\item{elcs}{a list of elcitation inputs}
}
\value{
An aggregated elcitation object
}
\description{
Pool elcitations data
}
