% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expert.R
\name{aggregate_experts}
\alias{aggregate_experts}
\alias{print.experts}
\title{Aggregate the elicitation data of many experts}
\usage{
aggregate_experts(experts)

\method{print}{experts}(experts, ...)
}
\arguments{
\item{experts}{a list of experts}
}
\description{
Aggregate the elicitation data of many experts
}
\examples{
ExpA = new_expert("A", "Triangle", c(54.2, 10, 150), "Triangle", c(2.3, 0, 5), "Triangle", c(5.1, 3, 8))
ExpB = new_expert("B", "Triangle", c(158.8, 30, 230), "Triangle", c(5.7, 2, 15), "Triangle", c(9.9, 7, 13))
ExpC = new_expert("C", "Triangle", c(204.4, 30, 410), "Triangle", c(7.1, 2, 10), "Triangle", c(3.5, 2, 6))

experts <- aggregate_experts(list(ExpA, ExpB, ExpC))
}
