% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bass_con.R
\name{generate_diffusion_continuous}
\alias{generate_diffusion_continuous}
\title{Generate a Bass Diffusion curve in continuous time approach}
\usage{
generate_diffusion_continuous(m, p, q, t_min = 0, t_max = 20, dt = 1,
  ...)
}
\arguments{
\item{m}{upper bound of the number of adoptions}

\item{p}{coefficient of innovation}

\item{q}{coefficient of imitation}

\item{t_min}{initial time point}

\item{t_max}{the end of time frame}

\item{dt}{observation interval}

\item{...}{arguments for deSolve::ode}
}
\value{
data.frame(Time, N, dN); N for the cumulative adoptions, dN for new adoptions
}
\description{
Generate a Bass Diffusion curve in continuous time approach
}
\examples{
dc <- generate_diffusion_continuous(140, 0.03, 0.41)
dc <- ts(dc[2:3], 0)
ts.plot(dc, col=c("green", "blue"))
legend("right", lty=1, legend=c("N", "dN"), col=c("green", "blue"))
}
