% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{generate_diffusion_curves}
\alias{generate_diffusion_curves}
\alias{generate_diffusion_curves.DCGen.single}
\alias{generate_diffusion_curves.DCGen.multi}
\title{Generate diffusion curves based expert parameters}
\usage{
generate_diffusion_curves(pars, t_min = 0, t_max = NULL, dt = 1,
  progress = NULL)

\method{generate_diffusion_curves}{DCGen.single}(pars, t_min = 0,
  t_max = NULL, dt = 1, progress = NULL)

\method{generate_diffusion_curves}{DCGen.multi}(pars, t_min = 0,
  t_max = NULL, dt = 1, progress = NULL)
}
\arguments{
\item{pars}{parameters sampled from expert inputs}

\item{t_min}{start time}

\item{t_max}{end time}

\item{dt}{showing time interval}

\item{progress}{shiny progress bar; NULL if not needed}
}
\description{
Generate diffusion curves based expert parameters
}
\examples{
ExpA = new_expert("A", "Triangle", c(54.2, 10, 150),
                  "Triangle", c(2.3, 0, 5), "Triangle", c(5.1, 3, 8))
ExpB = new_expert("B", "Triangle", c(158.8, 30, 230),
                  "Triangle", c(5.7, 2, 15), "Triangle", c(9.9, 7, 13))
ExpC = new_expert("C", "Triangle", c(204.4, 30, 410),
                  "Triangle", c(7.1, 2, 10), "Triangle", c(3.5, 2, 6))

experts <- aggregate_experts(list(ExpA, ExpB, ExpC))
pars <- rand_parameters(experts, 1000, method='mixture', type='continuous')
curves <- generate_diffusion_curves(pars, 0, 10)
}
