% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_input.R
\name{input_elicitation}
\alias{input_elicitation}
\alias{print.elc.input}
\title{Input elicitation data}
\usage{
input_elicitation(dist, pars)

\method{print}{elc.input}(elc, ...)
}
\arguments{
\item{dist}{distribution of the input among ["Gamma", "Normal", "LogNormal", "StudentT", "Triangle"]}

\item{pars}{a vector of parameters}

\item{elc}{an elicitation object to be print}

\item{...}{additional arguments passed to the integrator or to the methods}
}
\value{
An elc.input object
}
\description{
Input elicitation data
}
\examples{
elc <- input_elicitation("Gamma", c(0.1, 0.01))
elc
}
