% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_input.R
\name{input_lt}
\alias{input_lt}
\title{Input elicitation data as a log student t distribution}
\usage{
input_lt(pars)
}
\arguments{
\item{pars}{a vector of [log(mean), log(std), df]}
}
\value{
An elc.input object
}
\description{
Input elicitation data as a log student t distribution
}
\examples{
elc <- input_lt(c(0.1, 0.1))
elc
}
