% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_input.R
\name{input_norm}
\alias{input_norm}
\title{Input elicitation data as a normal distribution}
\usage{
input_norm(pars)
}
\arguments{
\item{pars}{a vector of [mean, std]}
}
\value{
An elc.input object
}
\description{
Input elicitation data as a normal distribution
}
\examples{
elc <- input_norm(c(0.1, 0.1))
elc
}
