% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_input.R
\name{input_t}
\alias{input_t}
\title{Input elicitation data as a student t distribution}
\usage{
input_t(pars)
}
\arguments{
\item{pars}{a vector of [mean, std, df]}
}
\value{
An elc.input object
}
\description{
Input elicitation data as a student t distribution
}
\examples{
elc <- input_t(c(0.1, 0.1))
elc
}
