% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_input.R
\name{input_triangle}
\alias{input_triangle}
\title{Input elicitation data as a triangle distribution}
\usage{
input_triangle(pars)
}
\arguments{
\item{pars}{a vector of [peak, min, max]}
}
\value{
An elc.input object
}
\description{
Input elicitation data as a triangle distribution
}
\examples{
elc <- input_triangle(c(10, 8, 12))
elc
}
