% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expert.R
\name{new_expert}
\alias{new_expert}
\alias{print.expert}
\title{Generate an expert with DCG input distributions}
\usage{
new_expert(name, dist_m, pars_m, dist_n1, pars_n1, dist_t, pars_t)

\method{print}{expert}(expert, ...)
}
\arguments{
\item{name}{Name of the expert}

\item{dist_m}{distribution of the cap of adoptions}

\item{pars_m}{parameters of m}

\item{dist_n1}{distribution of the adoptions in the first period}

\item{pars_n1}{parameters of N1}

\item{dist_t}{distribution of the time when new adoption starts declining (t')}

\item{pars_t}{parameters of t'}
}
\description{
Generate an expert with DCG input distributions
}
\examples{
Exp = new_expert("A", "Triangle", c(54.2, 10, 150), "Triangle", c(2.3, 0, 5), "Triangle", c(5.1, 3, 8))
print(Exp)
}
