% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bass_con.R
\name{pq2nt_continuous}
\alias{pq2nt_continuous}
\alias{nt2pq_continuous}
\title{Translation of parameters between (m, p , q) and (m, n1, t)}
\usage{
pq2nt_continuous(m, p, q)

nt2pq_continuous(m, n1, t, wt = c(0.5, 0.5))
}
\arguments{
\item{m}{upper bound of the number of adoptions}

\item{p}{coefficient of innovation}

\item{q}{coefficient of imitation}

\item{n1}{cumulative adoptions in the first period}

\item{t}{timing when new adoptions start decreasing}

\item{wt}{weight for n1 and t in finding (m, p, q)}
}
\value{
A list of translated parameters
}
\description{
Translation of parameters between (m, p , q) and (m, n1, t)
}
\examples{
pq2nt_continuous(140, 0.03, 0.41)
nt2pq_continuous(140, 5.94, 5.08)
}
