% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bass_dis.R
\name{pq2nt_discrete}
\alias{pq2nt_discrete}
\alias{nt2pq_discrete}
\title{Translation of parameters between (m, p, q) and (m, n1, t) using discrete time approach}
\usage{
pq2nt_discrete(m, p, q, dt = 0.01)

nt2pq_discrete(m, n1, t, wt = c(0.5, 0.5))
}
\arguments{
\item{m}{upper bound of the number of adoptions}

\item{p}{coefficient of innovation}

\item{q}{coefficient of imitation}

\item{dt}{precision of time}

\item{n1}{cumulative adoptions in the first period}

\item{t}{timing when new adoptions start decreasing}

\item{wt}{weight for n1 and t in finding (m, p, q)}
}
\value{
A list of translated parameters
}
\description{
Translation of parameters between (m, p, q) and (m, n1, t) using discrete time approach
}
\examples{
pq2nt_discrete(140, 0.03, 0.41)
nt2pq_discrete(140, 5.94, 5.08)
}
