% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_input.R
\name{print.elc.input}
\alias{print.elc.input}
\title{Print an elc.input object}
\usage{
\method{print}{elc.input}(elc)
}
\arguments{
\item{elc}{an elc.input object}
}
\description{
Print an elc.input object
}
