% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_rand.R
\name{rand_elcitation}
\alias{rand_elcitation}
\alias{rand_elcitation.elc.input}
\alias{rand_elcitation.elc.agg}
\title{Generate random values of an elicitation object}
\usage{
rand_elcitation(elc, n = 1, ...)

\method{rand_elcitation}{elc.input}(elc, n = 1)

\method{rand_elcitation}{elc.agg}(elc, n = 1, method = c("mixture",
  "average"))
}
\arguments{
\item{elc}{an elication object elc.input or elc.agg}

\item{n}{number of samples}

\item{...}{optional arguments}

\item{method}{"mixture" or "average" see below}
}
\value{
a vector of sample values
}
\description{
Generate random values of an elicitation object
}
\examples{
expert1 <- input_norm(c(10, 2))
expert2 <- input_gamma(c(0.1, 0.01))

# Single expert
summary(rand_elcitation(expert1, 100))

# Aggregated expert
experts <- aggregate_elcitations(list(expert1, expert2))
summary(rand_elcitation(experts, 100))
}
