% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expert.R
\name{rand_parameters}
\alias{rand_parameters}
\alias{rand_parameters.expert}
\alias{rand_parameters.experts}
\title{Generate random samples of diffusion curve parameters given elicitation inputs}
\usage{
rand_parameters(expert, n = 1, type = c("continuous", "discrete"), ...)

\method{rand_parameters}{expert}(expert, n = 100, max_try = 20,
  type = c("continuous", "discrete"), ...)

\method{rand_parameters}{experts}(expert, n = 100,
  method = c("mixture", "average"), type = c("continuous", "discrete"),
  max_try = 20, ...)
}
\arguments{
\item{type}{"continuous" or "discrete" time}

\item{...}{}

\item{max_try}{maximium tests if parameters are infeasible}

\item{method}{aggregation method of "mixture" or "average"}
}
\description{
Generate random samples of diffusion curve parameters given elicitation inputs
}
