% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{visualise_inputs}
\alias{visualise_inputs}
\alias{visualise_fitted}
\alias{visualise_comparing}
\alias{visualise_curves}
\title{Title}
\usage{
visualise_inputs(pars, col = "#FF6666")

visualise_fitted(pars, col = "#FF6666")

visualise_comparing(pars, col_i = "blue", col_o = "black")

visualise_curves(curves, dN = F, statistics = T, ci_range = 95,
  average = c("mean", "median"))
}
\arguments{
\item{pars}{sampled parameters}

\item{col}{colour of density diagrams}

\item{col_i}{colour of input variables}

\item{col_o}{colour of fitted variables}

\item{curves}{generated diffusion curves}

\item{dN}{show new adoption or not}

\item{statistics}{show statistics or curves}

\item{ci_range}{range of confidence intervals (0 - 100)}

\item{average}{average from ["mean", "median"]}
}
\description{
Title
}
