% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_input.R
\name{input_elcitation}
\alias{input_elcitation}
\title{Input elicitation data}
\usage{
input_elcitation(dist, pars)
}
\arguments{
\item{dist}{distribution of the input among ["Gamma", "Normal", "LogNormal", "StudentT", "Triangle"]}

\item{pars}{a vector of parameters}
}
\value{
An elc.input object
}
\description{
Input elicitation data
}
\examples{
elc <- input_elcitation("Gamma")
elc
}
