% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elc_input.R
\name{input_lnorm}
\alias{input_lnorm}
\title{Input elicitation data as a normal distribution}
\usage{
input_lnorm(pars)
}
\arguments{
\item{pars}{a vector of [log(mean), log(std)]}
}
\value{
An elc.input object
}
\description{
Input elicitation data as a normal distribution
}
\examples{
elc <- input_lnorm(c(2.5, 1))
elc
}
